Other way of splitting
